Update to 10.0.4

+++ front/computer.form.php
@@ -36,6 +36,7 @@
 use Glpi\Event;
 
 include('../inc/includes.php');
+
 Session::checkRight("computer", READ);
 
 if (!isset($_GET["id"])) {
@@ -45,31 +46,12 @@ if (!isset($_GET["id"])) {
 if (!isset($_GET["withtemplate"])) {
     $_GET["withtemplate"] = "";
 }
-$back = false;
+
 $computer = new Computer();
 //Add a new computer
-
-
 if (isset($_POST["add"])) {
-    //var_dump($_POST);
-    $computertypeid = $_POST['computertypes_id'];
-    $manufacturesid = $_POST['manufacturers_id'];
-    $modelid = $_POST['computermodels_id'];
-
-   
-    if($computertypeid == 0 || $manufacturesid == 0 || $modelid == 0){
-        echo '<script language="javascript">';
-        echo 'if(confirm("Erro ao adicionar computador\nOs campos Modelo, Fabricante e Tipo devem ser preenchidos.") == true){ history.go(-1)}';
-        //echo 'else{javascript:window.history.back();}';
-        echo '</script>';
-    }
-    else{
-
-    
     $computer->check(-1, CREATE, $_POST);
-    
- // array('baz' => ['end'])
-     if ($newID = $computer->add($_POST)) {
+    if ($newID = $computer->add($_POST)) {
         Event::log(
             $newID,
             "computers",
@@ -85,7 +67,7 @@ if (isset($_POST["add"])) {
     Html::back();
 
    // delete a computer
-}} else if (isset($_POST["delete"])) {
+} else if (isset($_POST["delete"])) {
     $computer->check($_POST['id'], DELETE);
     $ok = $computer->delete($_POST);
     if ($ok) {
@@ -129,19 +111,6 @@ if (isset($_POST["add"])) {
    //update a computer
 } else if (isset($_POST["update"])) {
     $computer->check($_POST['id'], UPDATE);
-    $computertypeid = $_POST['computertypes_id'];
-    $manufacturesid = $_POST['manufacturers_id'];
-    $modelid = $_POST['computermodels_id'];
-
-   
-    if($computertypeid == 0 || $manufacturesid == 0 || $modelid == 0){
-        echo '<script language="javascript">';
-        echo 'if(confirm("Erro ao atualizar computador\nOs campos Modelo, Fabricante e Tipo devem ser preenchidos.") == true){ history.go(-1)}';
-        //echo 'else{javascript:window.history.back();}';
-        echo '</script>';
-    }
-    else{
-
     $computer->update($_POST);
     Event::log(
         $_POST["id"],
@@ -151,9 +120,7 @@ if (isset($_POST["add"])) {
         //TRANS: %s is the user login
         sprintf(__('%s updates an item'), $_SESSION["glpiname"])
     );
-    html::Back();
-    }
-    
+    Html::back();
 
    // Disconnect a computer from a printer/monitor/phone/peripheral
 } else {//print computer information
