+++ front/phone.form.php
@@ -51,19 +51,6 @@ $phone = new Phone();
 if (isset($_POST["add"])) {
     $phone->check(-1, CREATE, $_POST);
 
-    $phonetypeid = $_POST['phonetypes_id'];
-    $manufacturesid = $_POST['manufacturers_id'];
-    $modelid = $_POST['phonemodels_id'];
-
-   
-    if($phonetypeid == 0 || $manufacturesid == 0 || $modelid == 0){
-        echo '<script language="javascript">';
-        echo 'if(confirm("Erro ao adicionar telefone!\nOs campos Modelo, Fabricante e Tipo devem ser preenchidos.") == true){ history.go(-1)}';
-        //echo 'else{javascript:window.history.back();}';
-        echo '</script>';
-    }
-    else {
-
     if ($newID = $phone->add($_POST)) {
         Event::log(
             $newID,
@@ -77,7 +64,7 @@ if (isset($_POST["add"])) {
         }
     }
     Html::back();
-} }else if (isset($_POST["delete"])) {
+} else if (isset($_POST["delete"])) {
     $phone->check($_POST["id"], DELETE);
     $phone->delete($_POST);
 
@@ -119,18 +106,6 @@ if (isset($_POST["add"])) {
 } else if (isset($_POST["update"])) {
     $phone->check($_POST["id"], UPDATE);
 
-    $phonetypeid = $_POST['phonetypes_id'];
-    $manufacturesid = $_POST['manufacturers_id'];
-    $modelid = $_POST['phonemodels_id'];
-
-   
-    if($phonetypeid == 0 || $manufacturesid == 0 || $modelid == 0){
-        echo '<script language="javascript">';
-        echo 'if(confirm("Erro ao adicionar telefone!\nOs campos Modelo, Fabricante e Tipo devem ser preenchidos.") == true){ history.go(-1)}';
-        //echo 'else{javascript:window.history.back();}';
-        echo '</script>';
-    }
-    else {
     $phone->update($_POST);
     Event::log(
         $_POST["id"],
@@ -141,7 +116,7 @@ if (isset($_POST["add"])) {
         sprintf(__('%s updates an item'), $_SESSION["glpiname"])
     );
     Html::back();
-}} else if (isset($_POST["unglobalize"])) {
+} else if (isset($_POST["unglobalize"])) {
     $phone->check($_POST["id"], UPDATE);
 
     Computer_Item::unglobalizeItem($phone);
